Refactored Admin Interface
Refactored format of admin interface. All of front-end visuals for user (ie prints, inputs) are implemented in the adminInterface.py file.

functions.py now mainly consists of just database queries.
